// server/src/enums/errors-code.enums.ts
export enum ErrorCodes {
	SUCCESS = 0,
	FAILED = 1,
	USER_NOT_FOUND = 2,
	EMAIL_NOT_FOUND = 3,
	ACCOUNT_ALREADY_VERIFIED = 4,
	INVALID_CREDENTIALS = 5,
	EMAIL_NOT_VERIFIED = 6,
	ACCESS_DENIED = 7,
	INVALID_TOKEN = 8,
	RESOURCE_NOT_FOUND = 9,
	INTERNAL_SERVER_ERROR = 10,
	UPDATED_FAILED = 11,
	INVALID_ID = 12,
	BAD_REQUEST = 13,
	INVALID_OTP = 14,
	UNDEFINED_ROUTE = 15,
	NOT_AUTH = 16,
	IS_EXPIRED = 17,
	USER_IS_ALREADY_EXISTED = 18,
	REQUEST_ERROR = 13,
}